Fix API breakage
Welcome to Golang, gentlemen.

Only a sticking plaster for now; comment out the (broken) setIcon code
because I don't know how it works, and use a true boolean variable
rather than an integer in the setColorKey call. Really, the original
developer needs to come out of hibernation to fix setIcon.